app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
minio==7.2.15
open-clip-torch==2.29.0
torch==2.6.0
Pillow==11.2.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
flake8==7.2.0
